requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
httpx==0.24.1
transformers==4.35.0
sentence-transformers==2.2.2
torch==2.9.1
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.0
requests==2.31.0
pinecone-client==3.0.3
redis==4.5.4
celery==5.2.7
SQLAlchemy==2.0.22
psycopg2-binary==2.9.10
structlog==23.3.0
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
sentry-sdk==1.26.0
sentencepiece==0.2.1
torchvision==0.24.1
tokenizers==0.14.1
safetensors==0.7.0
scipy==1.15.3
nltk==3.9.2